+++ src/main/java/com/caoccao/javet/interop/V8Runtime.java
@@ -775,6 +775,21 @@ public class V8Runtime implements IJavetClosable, IV8Creatable, IV8Convertible {
         return v8Native.set(handle, iV8ValueObject.getHandle(), iV8ValueObject.getType().getId(), key, value);
     }
 
+    public boolean setAccessor(
+            IV8ValueObject iV8ValueObject,
+            String propertyName,
+            JavetCallbackContext javetCallbackContextGetter,
+            JavetCallbackContext javetCallbackContextSetter) throws JavetException {
+        Objects.requireNonNull(javetCallbackContextGetter);
+        boolean isAccessorSet = v8Native.setAccessor(handle, iV8ValueObject.getHandle(), iV8ValueObject.getType().getId(),
+                propertyName, javetCallbackContextGetter, javetCallbackContextSetter);
+        callbackContextMap.put(javetCallbackContextGetter.getHandle(), javetCallbackContextGetter);
+        if (javetCallbackContextSetter != null) {
+            callbackContextMap.put(javetCallbackContextSetter.getHandle(), javetCallbackContextSetter);
+        }
+        return isAccessorSet;
+    }
+
     public boolean setProperty(IV8ValueObject iV8ValueObject, V8Value key, V8Value value) throws JavetException {
         decorateV8Values(key, value);
         return v8Native.setProperty(handle, iV8ValueObject.getHandle(), iV8ValueObject.getType().getId(), key, value);